[MPT-123] Updating file names
