+++ packages/GIS-ly/ui/src/model/GIS/Coordinate.spec.ts
@@ -1,5 +1,5 @@
 import { Coordinate, ICoordinate } from "./Coordinate";
-import { GeoJSONFeatureCollection } from "./GeoJSON/GeoJSON";
+import { GeoJSONFeatureCollection } from "./GeoJSON/GeoJSON_Refactor";
 
 describe('##Coordinate', () => {
