+++ packages/GIS-ly/ui/src/model/GIS/GeoJSON/Geometries/index.ts
@@ -1,8 +1,8 @@
+export { GeometryType as Geometry } from './Geometry';
+export * from './GeometryCollection';
 export * from './LineString';
 export * from './MultiLineString';
 export * from './MultiPoint';
 export * from './MultiPolygon';
 export * from './Point';
 export * from './Polygon';
-export * from './GeometryCollection';
-export { GeometryType as Geometry } from './Geometry';
