+++ packages/GIS-ly/ui/src/model/GIS/index.ts
@@ -1,3 +1,3 @@
-export * from './GeoJSON/GeoJSON';
+export * from './GeoJSON/GeoJSON_Refactor';
 export * from './Track';
 export * from './GeoJSON/BoundingBox';
